Fix type of dest_addr_len to respect return value of udp_set_url. Patch by Luca Abeni.
Originally committed as revision 13114 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavformat/udp.c
@@ -51,7 +51,7 @@ typedef struct {
 #else
     struct sockaddr_storage dest_addr;
 #endif
-    size_t dest_addr_len;
+    int dest_addr_len;
 } UDPContext;
 
 #define UDP_TX_BUF_SIZE 32768
